app/requirements.txt
pandas==2.3.2
matplotlib==3.10.5
polygon-api-client==1.15.3